+++ test/functional/feature_index_prune.py
@@ -5,13 +5,13 @@
 """Test indices in conjunction with prune."""
 import concurrent.futures
 import os
-from test_framework.authproxy import JSONRPCException
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.test_node import TestNode
 from test_framework.util import (
     assert_equal,
     assert_greater_than,
     assert_raises_rpc_error,
+    JSONRPCException,
 )
 
 from typing import List, Any
