+++ test/functional/test_framework/authproxy.py
@@ -44,17 +44,12 @@ import socket
 import time
 import urllib.parse
 
+from .util import JSONRPCException
+
 HTTP_TIMEOUT = 30
 USER_AGENT = "AuthServiceProxy/0.1"
 
 log = logging.getLogger("BitcoinRPC")
-
-class JSONRPCException(Exception):
-    def __init__(self, rpc_error, http_status=None):
-        super().__init__(f"{rpc_error} [http_status={http_status}]")
-        self.error = rpc_error
-        self.http_status = http_status
-
 
 def serialization_fallback(o):
     if isinstance(o, decimal.Decimal):
