+++ test/functional/test_framework/test_framework.py
@@ -22,7 +22,6 @@ import tempfile
 import time
 
 from .address import create_deterministic_address_bcrt1_p2tr_op_true
-from .authproxy import JSONRPCException
 from . import coverage
 from .p2p import NetworkThread
 from .test_node import TestNode
@@ -40,6 +39,7 @@ from .util import (
     p2p_port,
     wait_until_helper_internal,
     wallet_importprivkey,
+    JSONRPCException,
 )
