+++ test/functional/test_framework/util.py
@@ -20,7 +20,6 @@ import shlex
 import time
 import types
 
-from .authproxy import JSONRPCException
 from .descriptors import descsum_create
 from collections.abc import Callable
 from typing import Optional, Union
@@ -28,6 +27,12 @@ from typing import Optional, Union
 SATOSHI_PRECISION = Decimal('0.00000001')
 
 logger = logging.getLogger("TestFramework.utils")
+
+class JSONRPCException(Exception):
+    def __init__(self, rpc_error, http_status=None):
+        super().__init__(f"{rpc_error} [http_status={http_status}]")
+        self.error = rpc_error
+        self.http_status = http_status
 
 # Assert functions
 ##################
