+++ test/functional/wallet_multiwallet.py
@@ -12,7 +12,6 @@ import platform
 import shutil
 import stat
 
-from test_framework.authproxy import JSONRPCException
 from test_framework.blocktools import COINBASE_MATURITY
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.test_node import ErrorMatch
@@ -20,6 +19,7 @@ from test_framework.util import (
     assert_equal,
     assert_raises_rpc_error,
     ensure_for,
+    JSONRPCException,
 )
 
 got_loading_error = False
